requirements.txt
fastapi
uvicorn
biopython
requests
transformers
torch
python-dotenv
sentence-transformers
torch
